Renamed /settings/connection to /settings/connections
