+++ lib/Convos/Controller/User.pm
@@ -274,7 +274,7 @@ sub _register_html_conn_url_redirect {
     return $self->redirect_to($redirect_url);
   }
   else {
-    my $redirect_url = $self->url_for('/settings/connection');
+    my $redirect_url = $self->url_for('/settings/connections');
     $redirect_url->query->param(uri => $conn_url);
     return $self->redirect_to($redirect_url);
   }
